+++ lib/src/log.dart
@@ -1,11 +1,31 @@
+/// Function signature for custom log handlers
+typedef LogHandler = void Function(
+    String level, String channel, String message);
+
 var psdEnableLogging = false;
+LogHandler? _customLogHandler;
+
+/// Sets a custom log handler for PSD SDK logging
+///
+/// Example:
+/// ```dart
+/// PsdLogger.setLogHandler((level, channel, message) {
+///   print('[$level] $channel: $message');
+/// });
+/// ```
+void setLogHandler(LogHandler? handler) {
+  _customLogHandler = handler;
+}
 
 void psdWarning(List<String> args) {
   if (psdEnableLogging) {
     final channel = args[0];
     final msg = args.sublist(1).reduce((value, element) => '$value $element');
     setLastError(msg);
-    print('***WARNING*** [$channel] $msg');
+
+    if (_customLogHandler != null) {
+      _customLogHandler!('WARNING', channel, msg);
+    }
   }
 }
 
@@ -14,7 +34,10 @@ void psdError(List<String> args) {
     final channel = args[0];
     final msg = args.sublist(1).reduce((value, element) => '$value $element');
     setLastError(msg);
-    print('***ERROR*** [$channel] $msg');
+
+    if (_customLogHandler != null) {
+      _customLogHandler!('ERROR', channel, msg);
+    }
   }
 }